fix: putting MockBulletViewModel in its own file
